split each macro_attr into own file for each Co arg
