+++ src/stack/nightly_tests.rs
@@ -12,3 +12,18 @@ fn async_closure() {
     assert_eq!(gen.resume(), GeneratorState::Yielded(10));
     assert_eq!(gen.resume(), GeneratorState::Complete("done"));
 }
+
+#[test]
+fn stack__closure() {
+    let mut shelf = crate::Shelf::new();
+    #[yielder_cls(u8)]
+    let gen = unsafe {
+        Gen::new(&mut shelf, async move || {
+            let mut n = 1_u8;
+            while n < 10 {
+                yield_! { n };
+                n += 2;
+            }
+        })
+    };
+}
